Added old nest version (2.16) with changed Hebbian rule to the new branch

+++ nestkernel/music_event_handler.h
@@ -47,27 +47,27 @@ class MusicEventHandler : public MUSIC::EventHandlerGlobalIndex
 {
 public:
   MusicEventHandler();
-  MusicEventHandler( std::string portname, double acceptable_latency, int max_buffered );
+  MusicEventHandler( std::string portname,
+    double acceptable_latency,
+    int max_buffered );
 
   virtual ~MusicEventHandler();
 
   /**
    * Register a new node to a specific channel on this port.
    */
-  void register_channel( size_t channel, nest::Node* mp );
+  void register_channel( int channel, nest::Node* mp );
 
   /**
    * Publish the MUSIC port.
-   *
    * This method has to be called once before the first simulation to
    * tell MUSIC which channels lie on which processor.
    */
   void publish_port();
 
   /**
-   * Called by MUSIC from within tick() to deliver events to NEST.
-   *
-   * This function only queues the events. Delivery to the
+   * Called by MUSIC from within tick() to deliver events to
+   * NEST. This function only queues the events. Delivery to the
    * targets takes place in update().
    */
   void operator()( double t, MUSIC::GlobalIndex channel );
@@ -91,16 +91,17 @@ private:
   int max_buffered_;
 
   /**
-   * Buffers incoming spike events until they are due.
-   *
-   * The vector has one entry per channel. The priority queues used within the vector
+   * Buffers incoming spike events until they are due. The vector has
+   * one entry per channel. The priority queues used within the vector
    * implement min-heaps stored in vectors.
    */
-  std::vector< std::priority_queue< double, std::vector< double >, std::greater< double > > > eventqueue_;
+  std::vector< std::priority_queue< double,
+    std::vector< double >,
+    std::greater< double > > > eventqueue_;
 };
 
 } // namespace nest
 
-#endif /* #ifdef HAVE_MUSIC */
+#endif // HAVE_MUSIC
 
-#endif /* #ifndef MUSIC_EVENT_HANDLER */
+#endif // MUSIC_EVENT_HANDLER
